Updated openstack/openstack Project: openstack/ironic  1985fbb797399e6d4cc0475788e67c0075736d4e
Replace dict.get(key) in api & conductor tests

This replaces dict.get(key) with dict[key] in
tests/unit/api & tests/unit/conductor, since we should
be explicit about whether we expect the key to exist or not.

In particular, we shouldn't be doing
    self.assertIsNone(dict.get(key))
since it will pass if 1. the value is None or 2. if the key isn't
in the dict. For the case of 2, .assertIsNone() is replaced with
.assertNotIn().

Change-Id: Ia305a221389e39d759b0881620c200c625fc98e1
Closes-Bug: #1590808