Updated openstack/openstack Project: openstack-infra/project-config  fd5fa70a588483fbd8b5fc0455cd6bd84096cb0b
Upgrade experimental searchlight ES 2.3.2

Verification of Searchlight using ES 2.3.2 has been finalized. All
tests are enabled and the jobs are passing. It is time to remove these
jobs from experimental and add them to the regular check and gate jobs.

Modified the testing matrix to get the most bang for the buck in terms
of the different combinations that will be tests.

Before:
- Python 2.7/ElasticSearch 1.7.2
- Python 3.4/ElasticSearch 1.7.2

Now:
- Python 2.7/ElasticSearch 1.7.2
- Python 2.7/ElasticSearch 2.3.2
- Python 3.4/ElasticSearch 2.3.2

Change-Id: I43bc38d58d4039c1c7168d3a7c99157f3a3f0689
Partial-Bug: #1582316