Updated openstack/openstack Project: openstack/cinder  4c7cdff9abef4a22ddba251f6ebd239dd3844a02
Call _init_volume_driver in backup manager

When fixing backup using temp snapshot code path, the call to
_init_volume_driver was missed:
    https://review.openstack.org/#/c/321943/

For drivers that rely on do_setup to initialize the driver, this
is a problem because the driver cannot be used without this call.

This patch fixes the problem by calling _init_volume_driver in
init_host.

Change-Id: Ic197046a480dbf8d73a229580e59cd6d489355fa
Closes-Bug: #1594225