Updated openstack/openstack Project: openstack/neutron  70d2fc8cc19c9b2fcdce108304dd1e76959bed7a
Check for RetryRequest in MultipleException types

The is_retriable decorator that examines MultipleException
exceptions to determine if they should be retried was not
checking for RetryRequest exceptions. So if a RetryRequest
was raised in a callback, it would not have been retried.

This adds RetryRequest to the list of exceptions to look for.

Closes-Bug: #1595345
Change-Id: Iec679a622bbdc0c15bddd1df0adf758f1f1e8798