Updated openstack/openstack Project: openstack/nova  5730a39ed79a4b1d35d230fc9d4ba239481c9532
fix developer docs on API

During the defcore discussion about the fact that the extensions
facility was deleted from Nova, folks pointed to this doc and the word
extensions to state that it was clearly still supported. There are
many levels of confusion there.

This tries to crisp up the introduction and links to the API docs for
Nova.

Change-Id: Ic0be29fb14bfb00ed7691439318325967d138dc9