Updated openstack/openstack Project: openstack-infra/elastic-recheck  3706436fd7d669f9483a7ebefdf9d1ec729710aa
Update query for ceph scheduling bug 1475333

The previous query wasn't hitting on the ceph job
at all, this is a more accurate query for the ceph
job bug. There are only 7 hits in 7 days in the check
queue but it's on multiple changes on the same ceph
job, so it's more accurate.

Change-Id: Idc888ebb4daf61bbbe3077c0040ff0bc5fa5a9df
Related-Bug: #1475333