Updated openstack/openstack Project: openstack/nova  f3700e20d941c349f7ad3590a56c6a663dcce526
libvirt: split out code for detecting live migration job type

When we see a job type of "NONE", we have code to convert that
to another more useful job type. Split this code out into its
own method to make it easier to test it in isolation. This new
testing exposed a bug in log debug messages where the 'ex'
parameter was not being correctly passed in a dict.

Refactoring of mocking in migration tests in test_driver.py is
left as a future exercise once this whole series is merged.

Implements: blueprint libvirt-clean-driver
Change-Id: I80e5b2c22a03556bfce29a74ae304cef6b6841cb