Updated openstack/openstack Project: openstack-infra/system-config  18f451fec78f77ac7fd4eaeab3d4606b25cd2f31
Add #openstack-pkg to statusbot & meetbot_channels

This patch adds IRC bots to #openstack-pkg, for the packaging-deb group.

It looks like "openstackgerrit" (gerrit change reports) is already there,
but "openstack" (channel logging) and "openstackstatus" (infra status
notifications, success submissions) are not yet because the channel is
not in the "meetbot_channels" and "statusbot_channels" sections.

Change-Id: I91adbf4392cb028de55693f5c09f438cd4d9fb54