Updated openstack/openstack Project: openstack-infra/project-config  8d38ec40c62875b810ea798135503bb9516a9276
Don't run grenade on liberty

This moves the grenade jobs to not run on liberty (as kilo is no
longer a thing). The multinode special casing can be dropped because
it was working as of mitaka for nova-net and neutron, and the base
grenade definition does not include a $, so matches these.

Change-Id: I60c834c9a20e4a1eb10a24b2d0cf0080c65c399e