Updated openstack/openstack Project: openstack/nova  aa963b3136f3c0ce0d96dd790a993879e3a7ac6e
collapse servers.ViewBuilderV21 into servers.ViewBuilder

We used to have 2 different server views to handle the differences
between base V20 and V21 APIs, however with the drop of the legacy V20
stack we only need one. This prevents the duplication in show method,
and simplifies the code.

The diff looks a little odd because this was copying the v21 show over
the v20 show, so the unrelated whitespace changes were actually in the
old method, and do make it more readable.

Part of bp:api-no-more-extensions

Change-Id: I8a03ba4b896a16d97131d51761760622520d01fd