Updated openstack/openstack Project: openstack/diskimage-builder  89befa1fb11f68e8ebba04a2966a0bd402e70f9e
Handle locales install on Fedora 24

Fedora 24 has split locales into separate packages.  Testing revealed
what is possibly a bug in the choosing of default packages, so add a
small work-around to ensure the minimal locale pack is installed.

This appears to be the only change required for fedora-minimal with
Fedora 24; at least to build with the project-config infra elements.

Change-Id: I64438c34c572ed96211384ae1bfb45b2949e4318