Updated openstack/openstack Project: openstack/nova  76e857cade0fe005d6ed6a22e88a412eb2e11b65
libvirt: virtlogd: use "log" element in char devices

This change introduces the usage of the <log> domain XML element for char
devices. A follow up change will use this object to store the boot messages
of an instance when the parent char device uses the virtlogd as backend
for its stdout/stderr output.

blueprint libvirt-virtlogd

Change-Id: I3fc83cfe20dc5facea2aff3e7a7dc795952999f4