Updated openstack/openstack Project: openstack/python-ironicclient  acc91e4e5033fd5c5b8b0bc1b78dec15dcd47573
Fix error returned by ironic --json node-validate

The option --json used the attribute "_info" to print the JSON
response from Ironic API. Some objects do not have this attribute,
which results in an error. This patch stops relying on "_info" and
uses the fields names to build the JSON array.

Change-Id: I6c0ff73a38cda21e147b3216d0c0d0539cc867bb
Closes-Bug: #1558688