Updated openstack/openstack Project: openstack/nova  2a39be0e1580683a4e847bf6f2b0bd1eacadb6d9
policy: Replaces 'authorize' in nova-api (part 2)

Adds 'fatal' argument (with default value True) to context.
If the argument is False, if a nova.exception.Forbidden
is raised, False is returned instead. This is the equivalent of
soft_authorize.

Adds docstring to nova.context.RequestContext.can.

Adds unit tests for nova.context.RequestContext.can.

Partially-Implements: bp policy-in-code

Change-Id: Ib6388822da3d3eb9b8fbd9b5c341cd8efeb790fe