Updated openstack/openstack Project: openstack-infra/storyboard  221683c213780cebdd9463e832c2e6aa163f7d40
Store history of a comment when editing

At the moment it is possible to edit comments via the API, but this
causes the previous content of the comment to be permanently lost.

This commit fixes this problem by keeping a record of the previous version
of a comment each time it is edited.

Change-Id: I0bcf4ba3555c8d96e91497742f424056e3123312