Updated openstack/openstack Project: openstack/openstack-manuals  8a2467162f32c900df4f2af7195315934a4f4084
New key manager config update

A configuration guide update for nova.conf and cinder.conf is needed
due to a new key manager interface. The interface is based on the
Castellan key manager which provides a consistent front-end interface
to support different back end key managers (Barbican is the default
back-end.) The code patch for Nova is merged, and corresponding patch
for Cinder is in review.

Depends-On: Ib563b0ea4b8b4bc1833bf52bf49a68546c384996
Depends-On: Ief8885bb4ca8d62b03cf1a52c25dd0e62c835bfe
Closes-Bug: #1592026

Change-Id: I99bcaf014c210f2598c1809b52ea46967b025a74