Updated openstack/openstack Project: openstack/horizon  9dd311abb2f7e41bb407a750c7520525932d377a
Addition of metadata for snapshot in VolumeSnapshot object

When accessing the snapshot object in Horizon codebase, we do not have
access to the metadata of the snapshot. This is because the
VolumeSnapshot object does not have metadata as a member. So added
'metadata' to the VolumeSnapshot object. The fix should provide access
to metadata.
Added test data and test cases for snapshot object with metadata.

Closes-Bug: 1542468

Change-Id: I4ae2e7a74d3ddbff8493b42f8f86c7002229cf8c