Updated openstack/openstack Project: openstack/nova  7a51a5ca469c5f57c1723f47938bcccde57cebdf
make failures on api_samples more clear

This changes the failure mode for api_samples in 2 ways. The first of
which is it pretty prints the 2 documents that are compared whenever
there is a failure. This makes finding the differences by eye a lot
easier.

Secondly, it adds a short circuit when comparing lists, if they are
both of length 1. The list comparison is weird, and you get very
unhelpful error messages if you are comparing 2 lists of length 1 and
one has extra keys. It doesn't help all situations, but this is pretty
common, so will make these a ton easier to figure out what's wrong.

All of this is extremely helpful in figuring out what you did wrong in
decomposing an API extension. And thus is part of

blueprint api-no-more-extensions

Change-Id: I5f487c6798485a58cc020723dd5e880720c51bed