Updated openstack/openstack Project: openstack/ironic  95ec6ab8b75641db98af19e6b76a481116ad6ad2
Fail early if ramdisk type is dib, and not building

Currently if IRONIC_BUILD_DEPLOY_RAMDISK is False, and ramdisk type is
DIB, we won't fail until the script attempts to run wget using an empty
address, this results in a hard to debug error. This patch ensures that
if we are not building the ramdisk but the ramdisk type is DIB that the
whole script fails early with an error message describing how to correct
the problem.

Change-Id: Ib9f31a3356ed64631129e5880e2657b0cf2014d2