Updated openstack/openstack Project: openstack/horizon  fa28acf97cb1a67af9166c9f853f7b903c41af0c
Place Create Image in global actions, not batch actions

See bug report for details.  Nothing should appear differently due to this
change in Horizon, but the effect is that anywhere we want to see global actions,
this will show up in the list.

Added test for registration and fixed over-general app.core reference.

Change-Id: I1dfd444e6f2dac6b5f8175eed336d3331f0f5a5a
Closes-Bug: 1596560