Updated openstack/openstack Project: openstack/diskimage-builder  3d85df15189e7fac4e4809219d1677fe8c37faf9
Fix sphinx-build to not depend on diskimage-builder

Currently, running sphinx_build fails for us because we depend on
diskimage-builder in our sphinx conf.py. This causes doc generation
on sites like rtfd to fail unless they install the diskimage-builder
module beforehand. We can, alternatively, import pbr directly and not
require the module as part of doc generation.

Change-Id: I41f222ff9c67950fc30841935a6a603f5718395e