Updated openstack/openstack Project: openstack/nova  35eaa27a04612d85c20f067ab5bf1425b4a4f78e
Don't check for file type in _find_base_file

Firstly, _find_base_file was differentiating between 2 different types
of 'non-original', but the caller was then treating them identically.
This differentiation was obviously redundant.

However, the caller was using them to populate originals. Note that
_scan_base_files has been called previously, and therefore originals
is already correctly populated. This means that _find_base_file does
not need to identify 'originals' at all.

Additionally, originals is only tracked at all so that 'originals' and
'non-originals' can be aged with different frequencies. This doesn't
really make any sense anyway, so this is a small step towards removing
a redundant config variable.

Change-Id: I893cfe23b48ee135ac6c14870d50d7f734c25247