Updated openstack/openstack Project: openstack/sahara  f48eb7d9855b06d99b744ccc33a3d1cc287c6004
Remove unecessary decorators from private interface

This commit removes the usage of the services decorators from the
tempest plugin tests. That decorator is an internal interface and
not meant for external consumption, and is now causing failures
because of recent tempest changes. Plugins should only use documented
stable interfaces. [1]

Also the use of the decorator here actually adds nothing to the tests.
They do 2 things, add a test attr to filter by service and add a skip
check on the service availability. However both of these are already
done, data_processing is already part of the test_id so filtering
will just work and the skip check for sahara is baked into the base
test class. So using this decorator does nothing except break the
plugin.

[1] http://docs.openstack.org/developer/tempest/plugin.html#stable-tempest-apis-plugins-may-use

Change-Id: I45587d065c1b7b7f28feb29e8845f8f0f8405b8c