Updated openstack/openstack Project: openstack/nova  293143b1a52aa39bf4cdf5156ed2c1deefdf64b4
libvirt: split off code for updating migration stats in the DB

Periodically during migration we need to update the migration
object with fine details of the progress, while also updating
the coarse instance progress field. Move this update logic
into a separate method to allow for testing it in isolation.

Implements: blueprint libvirt-clean-driver
Change-Id: Ic59cf00283b5efc680612bc719efa2c61fc1b01a