Updated openstack/openstack Project: openstack/nova  bbaf858097389354887bc7e7e8b07c0f4545b78e
libvirt: split out code for recovering after migration tasks

While migration jobs are running it is possible to have
tasks requested against the VM, such as pausing it. If
migration fails or is cancelled we need to do some recovery
actions for any tasks that were run. This task recovery
processing is handled by a separate inline method.
Move this inline method out to be a static method in
the class, and add explicit testing of it. This also
adds an explicit warning if it finds an unknown task.

Implements: blueprint libvirt-clean-driver
Change-Id: I676e05ba61657d350ac3c59056a0eeebcd5b7040