Updated openstack/openstack Project: openstack/nova  7890d91eed35bd513edd309e1e358eb14511be4f
Add new configuration option to turn postcopy on/off

Since QEMU 2.5 and Libvirt 1.3.3 it is possible to use postcopy
live migration to ensure an upper limit in the amount of memory
to be transferred duing the live migration process. This option
enables to live migrate instances with memory-intensive workloads
by driving the migration process by the destination host, i.e.,
after the switch to postcopy mode, the destination VM is the active
one and therefore any dirty pages are directly generated at the
destination hosts.

Partially-implements: blueprint auto-live-migration-completion

Change-Id: I35b8400604eece5ddfba3e35015bcd3500d9f605
Signed-off-by: Luis Tomas <luis5tb@gmail.com>
Co-Authored-By: Pawel Koniszewski <pawel.koniszewski@intel.com>