Updated openstack/openstack Project: openstack-infra/project-config  9e7e087de3e788bed4a597503ec4f8983c3d4c04
Updates to bindep fallback for Xenial

We assumed a precise/trusty split in the bindep fallback file in many
places, but now we need to handle precise, trusty, and xenial. Correct
this by defaulting to using just trusty on ubuntu packages and
explicitly exclude precise where it is not needed or will not function.

Change-Id: Ib44fd82f282d8b9115fc25dcf77e6b0676a886a3