Updated openstack/openstack Project: openstack/tempest  65abcb0ebb9df1589ffbdb6d7bb820b8bcc74aa4
Prepare the Manager class for tempest.lib

Remove CONF dependencies from the client manager base class,
to make it useful to external consumers (such as plugins).

The ultimate target is to have a manager which can be used with
as little as possible setup, which only instantiates the clients
which are actually needed by the test, and which allows to register
new service clients defined in plugins.

Since plugins already import both manager.Manager and
clients.Manager, we maintain for now both classes with their names.
The plan is to migrate the 6 core service client groups to
manager.Manager, so that those clients are available to all tests
along with plugin clients. That requires a few steps. I'm doing
changes in clients.Manager for now so that it's easier to review.

The result of this first step is:
- manager.Manager is moved to manager_lib.Manager and does not
  depend on CONF anymore, nor on any tempest unstable class.
  It does not provide any client yet.
  Add unit test coverage for this class.
- manager.Manager is still provided with backward compatible
  interface for plugins benefit.

Change-Id: Ic9ccc7037d15cdd4c6f1749eaeda13d4e7ee0114
Partially-implements: bp client-manager-refactor