Updated openstack/openstack Project: openstack/nova  19037640ea35e98cf073c21263f82d10b52e790b
Libvirt driver implementation of device tagging

A cloud user boots a VM with two network interfaces. One is connected
to a private network, the other to the public Internet. There is
currently no direct way to tell which interface is which. This patch
series proposes an optional virtual network interface tag that is set
by the user at boot time and is exposed in the metadata.

This patch introduces the libvirt driver implementation of
get_device_metadata.

Partially implements: blueprint virt-device-role-tagging
Co-authored-by: Artom Lifshitz <alifshit@redhat.com>
Change-Id: I5eadd10449aa06c77918452a0f65e69eecc03746