Updated openstack/openstack Project: openstack/nova  7b6a6805369b53cfaeae9659c5b09780210e216b
Detach volume after deleting instance with no host

If an instance is booted from a volume, shelved, and goes into an error
state due to some reason, the volume from which instance is booted
remains even the instance is deleted because instance has no host
associated with it.

Called _local_delete() to detach volume and destroy bdm if instance is
in shelved_offloaded state or has no host associated with it. This will
cleanup both volumes and the networks.

Note:
Ankit had submitted same patch [1] earlier which was reverted [2] due
to a race condition on jenkins if an instance is deleted when it is in
building state.  The patch was then rebumitted [3] fixing the
the failure of race condition by reverting the ObjectActionError
exception handling in _delete.  This patch was later re-reverted [4]
due to continued jenkins race conditions.

The current patch avoids the jenkins race condition by leaving the flow
for instances in the BUILDING state unchanged and only calling
_local_delete() on instances in the shelved_offloaded or error states
when the instance has no host associated with it.  This addresses the
concerns of the referenced bugs.

[1] Ic630ae7d026a9697afec46ac9ea40aea0f5b5ffb
[2] Id4e405e7579530ed1c1f22ccc972d45b6d185f41
[3] Ic107d8edc7ee7a4ebb04eac58ef0cdbf506d6173
[4] Ibcbe35b5d329b183c4d0e8233e8ada26ebc512c2

Co-Authored-By: Ankit Agrawal <ankit11.agrawal@nttdata.com>

Closes-Bug: 1404867
Closes-Bug: 1408527

Change-Id: I928a397c75b857e94bf5c002e50ec43a2bed9848