Updated openstack/openstack Project: openstack/nova  317f7d758cb7cfc3bf102b95919c2fd3d9b2a666
Add policy sample generation

This adds the entry point and config file necessary for using the
oslo.policy sample generation script. It also adds a tox target to
simplify the usage of it.

The command is also run in the docs tox target since the sample file
can be considered documentation.

Change-Id: I18a1fa567fc04b0e0a37c02f20024b2ce483f060
Partially-Implements: bp policy-in-code