Updated openstack/openstack Project: openstack/nova  59e6d617c72bf26f0020bff975d67bb4bbb82b34
Add automatic switching to postcopy mode when migration is not progressing

Since QEMU 2.5 and Libvirt 1.3.3 it is possible to use postcopy
live migration to ensure an upper limit in the amount of memory
to be transferred duing the live migration process. This option
enables to live migrate instances with memory-intensive workloads
by driving the migration process by the destination host, i.e.,
after the switch to postcopy mode, the destination VM is the active
one and therefore any dirty pages are directly generated at the
destination hosts. This patch ensures that the migration is switched
to postcopy mode if the migration process is not progressing more
than 10% each iteration due to memory dirtying ratio being faster
than page transferring to the destination.

Partially-implements: blueprint auto-live-migration-completion

Change-Id: If267fb7066e8323303bcc0f7e8f36513ef4d66b3
Signed-off-by: Luis Tomas <luis5tb@gmail.com>
Co-Authored-By: Pawel Koniszewski <pawel.koniszewski@intel.com>